commit with jwt

+++ resources/item.py
@@ -1,8 +1,15 @@
-from flask_jwt import JWT, jwt_required
+from flask_jwt_extended import (
+    Jwt_required, 
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from flask_restful import Resource, reqparse
 import sqlite3
 from flask import jsonify
 from code.models.item import ItemModel
+
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument(
@@ -18,7 +25,7 @@ class Item(Resource):
         help='Every item needs a store id'
         )
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item=ItemModel.find_by_name(name)
 
@@ -27,7 +34,7 @@ class Item(Resource):
         else:
             return {"message":"item not found"},404
 
-
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'An item with name: {} already exists'.format(name)},400
@@ -42,7 +49,11 @@ class Item(Resource):
         
         return item.json(), 201
 
+    @jwt_required
     def delete(self,name):
+        claims=get_jwt_claims()
+        if not claims["isAdmin"]:
+            return {"message":"Admin privilege required"},401
         item=ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -63,5 +74,13 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items':[item.json() for item in ItemModel.find_all()]}
+        user_id=get_jwt_identity()
+        items=[item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items":items}
+        return {"items":[item["name"] for item in items],
+                "message":"Login to see more data"
+
+        }
